{{ cookiecutter.project_slug }}/requirements/developer.txt
# Install the main set of requirements
-r ./base.txt

# Additional requirements useful for developers of {{ cookiecutter.project_name }}

# Django debugging
django-debug-toolbar
